Patch entry return to allow non-reg-updated display
The previous iteration required a meet_id to be passed in to user and
org registration data pull calls, which limited display to only data
that was pushed for the current meet. That's contrary to the
requirements: that any entry for the season be used for the whole season
unless changed.

Now, the latest entry (user or org) created after the season break is
used.