Password reset permits multipl clicks on same link
Previously, a reset password link click would reset a user's password to
something random, which would invalidate the use of the reset link. This
seemed like preferred behavior, but for some users (mobile phones
probably), every reset password link click would somehow become a
double-click (or rather, a twice clicked link), which would give the
user an error message instead of desired behavior.

This change causes the underlying password to not be reset, but it will
login the user from the reset link. It's no less secure except if the
user then doesn't change their password as instructed. And even then,
not a super big deal since in theory the reset link is unique and
unguessable.